added swagger and any bugs fixed

+++ src/cars/cars.service.ts
@@ -1,20 +1,24 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common'
 import { PrismaService } from '../prisma/prisma.service'
 
-import { CarsDto } from './dto/cars.dto';
+import { CarsDto } from './dto/cars.dto'
 
 @Injectable()
 export class CarsService {
   constructor(private prisma: PrismaService) {}
   async getAll() {
-    return await this.prisma.cars.findMany();
+    return await this.prisma.cars.findMany()
   }
   async getUnique(id: number) {
     return this.prisma.cars.findUnique({
       where: {
         id: Number(id),
       },
-    });
+    })
   }
   async create(dto: CarsDto) {
     return await this.prisma.cars.create({
@@ -24,9 +28,21 @@ export class CarsService {
         year: dto.year,
         car_price: dto.car_price,
       },
-    });
+    })
   }
   async update(id: number, dto: CarsDto) {
+    const existingCar = await this.prisma.cars.findUnique({
+      where: {
+        id: Number(id),
+      },
+    })
+
+    if (!existingCar) {
+      // Ma'lumot topilmadi, 404 Not Found qaytarish
+      throw new NotFoundException(`Car with id ${id} not found`)
+    }
+
+    // Ma'lumot topildi, yangilashni davom ettiramiz
     return await this.prisma.cars.update({
       where: {
         id: Number(id),
@@ -37,13 +53,24 @@ export class CarsService {
         year: dto.year,
         car_price: dto.car_price,
       },
-    });
+    })
   }
+
   async delete(id: number) {
-    await this.prisma.cars.delete({
+    const existingCar = await this.prisma.cars.findUnique({
       where: {
         id: Number(id),
       },
-    });
+    })
+    if (!existingCar) {
+      throw new NotFoundException(`Car with id ${id} not found`)
+    } else {
+      await this.prisma.cars.delete({
+        where: {
+          id: Number(id),
+        },
+      })
+      return 'Car deleted'
+    }
   }
 }
